+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.9.5 (not yet released)
 
+- added `OfflineManagementClient.serverVersion()` (server version discovery
+  in offline finally implemented)
+
 ## 0.9.4
 
 - management version 4.1.0 is now recognized